Added new register comment and route and linking from navbar
