+++ client/src/App.js
@@ -5,6 +5,7 @@ import CreateRecipe from "./components/recipes/create";
 import ShowRecipes from "./components/recipes/index";
 import ShowIngredients from "./components/ingredients/index";
 import Logon from "./components/auth/create";
+import Register from "./components/auth/register";
 import Navigation from "./components/navigation"
 
 import "bootstrap/dist/css/bootstrap.css";
@@ -20,6 +21,7 @@ const App = () => {
         <Route path="/recipe/" element={<ShowRecipes />} />
         <Route path="/ingredient/" element={<ShowIngredients />} />
         <Route path="/auth/" element={<Logon />} />
+        <Route path="/auth/register" element={<Register />} />
       </Routes>
     </div>
     </>
