+++ client/src/components/navigation.js
@@ -20,7 +20,6 @@ export default function Navigation() {
         <Navbar.Toggle aria-controls="basic-navbar-nav" />
         <Navbar.Collapse id="basic-navbar-nav">
           <Nav className="me-auto">
-            <Nav.Link href="/auth">Login</Nav.Link>
             <Nav.Link onClick={handleClick}>Log Out</Nav.Link>
             <NavDropdown title="Get Cooking" id="basic-nav-dropdown">
               <NavDropdown.Item href="/recipe">My Recipes</NavDropdown.Item>
@@ -34,10 +33,12 @@ export default function Navigation() {
         </Navbar.Collapse>
         <Navbar.Collapse id="basic-navbar-nav">
         <Nav className="ms-auto">
-        <Navbar.Text >
+          <Navbar.Text >
               Signed in as: <a href="/auth">Add Name Here</a>
-            </Navbar.Text>
-            </Nav>
+           </Navbar.Text>
+           <Nav.Link href="/auth">Login</Nav.Link>
+           <Nav.Link href="/auth/register">Register</Nav.Link>
+        </Nav>
         </Navbar.Collapse>
 
       </Container>
